+++ src/Blitz.Avalonia.Controls/GotoEditorImageConverter.cs
@@ -45,6 +45,7 @@ public class GotoEditorImageConverter : IMultiValueConverter
         {"notepad++.exe", "Notepad_plus_plus.png"},
         {"Code.cmd", "vscode.png"},
         {"cursor.cmd", "cursor.png"},
+        {"windsurf.cmd", "windsurf-logo.png"},
         {"uedit64.exe", "uelogo.png"},
         {"fleet.exe", "Fleet.png"},
         {"devenv.exe", "VisualStudio2022PNG.png"},