+++ src/Blitz.Goto/CodeExecuteNames.cs
@@ -7,6 +7,7 @@ public class CodeExecuteNames
 {
     public const string VSCode = "VsCodeGoto"; 
     public const string Cursor = "CursorGoto"; 
+    public const string Windsurf = "WindsurfGoto";
     public const string VisualStudio = "VisualStudioPlugin"; 
     public const string BlitzEdit = "BlitzEdit"; 
     public const string SublimeText = "SublimeText";